One more step required to change limits in Debian
In Debian 8 (Jessie) and everything based on it we need  an additional step in order to change limits for an user, not just editing /etc/security/limits.conf file.

Send error message to clients that produce invalid transactions

This replaces the raw json response from bitcoind with a more user-friendly
message.

Send error message to clients that produce invalid transactions

This replaces the raw json response from bitcoind with a more user-friendly
message.