Populate version from current git commit hash
This combines the convenience of not having to manually bump
version numbers with the ability to find out if a specific server
provides a specific feature or uses a specific bugfix.

This also adds an `electrum-server version` command to check the
currently running version.

This doesn't set the SHA-hash as a version for the installation process
because setuptools/distutils requires an incrementing version format.

The protocol version was moved to a different variable that is only used
in IRC peer discovery.